Added releaseDates GET
Gets ALL IDs, and requests ~150 games for the genre/release date. 62 Are
succesfully returned and saved in the genres and releaseDates arrays.

Also added definedNotWorkingIDs, which never return anything so we don't
request their information anymore. This list should be expanded in the
future to prevent even more useless requests.